\name{evalMedian}
\alias{evalMedian}
\title{
Pruning a classification tree using the total misclassification cost
}
\description{
This function is not invoked directly by the user but is used for its effects in the pruning procedure. See Galimberti \emph{et al.} (2012) for further details. 
}

\author{Giuliano Galimberti, Gabriele Soffritti, Matteo Di Maso

}


\seealso{
\code{\link[rpartScore:rpartScore]{rpartScore}}
}
\references{
Galimberti G., Soffritti G., Di Maso M. 2012 Classification trees for ordinal responses in \R: the \code{rpartScore} package. 
\emph{Journal of Statistical Software}. Submitted. 

}