\name{splitQuad}
\alias{splitQuad}
\title{
Generalized Gini splitting function  based on quadratic differences in scores
}
\description{
This function is not invoked directly by the user but is used for its effects in the tree growing procedure. See Galimberti \emph{et al.} (2012) for further details.

}

\author{Giuliano Galimberti, Gabriele Soffritti, Matteo Di Maso

}


\seealso{
\code{\link[rpartScore:rpartScore]{rpartScore}}
}
\references{
Galimberti G., Soffritti G., Di Maso M. 2012 Classification Trees for Ordinal Responses in \R: The \code{rpartScore} Package.
\emph{Journal of Statistical Software}, \bold{47}(10), 1-25.
\doi{10.18637/jss.v047.i10}.

}
