\name{initScore}
\alias{initScore}

\title{
Summarizing and visualizing classification trees
}
\description{
This function is not invoked directly by the user but is used for summarizing and visualizing a classification tree. See Galimberti \emph{et al.} (2012) for further details. 
 
}

\author{Giuliano Galimberti, Gabriele Soffritti, Matteo Di Maso

}


\seealso{
\code{\link[rpartScore:rpartScore]{rpartScore}}
}
\references{
Galimberti G., Soffritti G., Di Maso M. 2012 Classification trees for ordinal responses in \R: the \code{rpartScore} package. 
\emph{Journal of Statistical Software}. Submitted. 

}